+++ src/winstan.ts
@@ -1,5 +1,6 @@
 export { log } from './logger';
 export { setColors } from './conf/color';
+export { setColorize } from './format/color';
 export { setLevel } from './conf/level';
 export { setLocale } from './conf/locale';
 export { setService } from './conf/service';